Ram's fixes (#5)
* Create main.yml

* Update main.yml

* Update main.yml

* Update main.yml

* Minor fixes and reformatting.

* Update requirements.txt

* Remove main.yml

* cards.py, players.py, trick.py: minor changes supporting duplication of objects
copy function for future recreation of a state that can be mutable (for search tree)

compare_agents.py: added 2 functions to run alfa beta vs. single action agents.
it is not generic and should be modified.

game.py: changed to support the state and.
default arguments added to support future comarison of db games (not fully done yet)
# todo modify self.players to get a specific hand

match.py: updated according to object Game for creating a game from db (not fully done yet)

multi_agents.py: all the agents created by ram, modified from Class to function representation.
alfa-beta-search agent is added with a single evaluation cost function.
an idea of function that compares the scores to required target score is set, but is yet supported fully in the search.

state.py: a state of a game. it has no abstract class. should serve the MCTS too, as MCTS will use nodes on it's own that will hold a state of a game.

files i don't need: interfaces.py, agents.py

* Reformatting

* Clearing trick moved to state.apply_action

* Renaming agents

* Renaming agents

* Rephrased TODOs

* Rephrased TODOs

* minor fixes

* minor fixes

* minor fixes

* added TODO

* Removed agents.py

* Refactor names in compare_agents.py

Co-authored-by: ramf <ramf@cs.huji.ac.il>
Co-authored-by: marchellyn <marynar@cs.huji.ac.il>
Co-authored-by: oriyanh <oriyanh@cs.huji.ac.il>